merlin: rootdir: Tune zram performance
-Switch to LZ4 for better overall performance -Set max_comp_streams to
num of cores since upstream also moves this to# Please enter the commit
message for your changes. Lines starting percpu.# with '#' will be
ignored, and an empty message aborts the commit. -Set page-cluster to 0
as the incremental cost of reading 1 page at a# time is negligible for
zram# On branch 11
Bug: 38249616#
Test: boot and run zram-perf showing better performance# Changes to be committed: